Hibernate - test dodawania rekordu do tabeli
