--- src/connection/Hibernate_Controller.java
@@ -1,36 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package connection;
-
-import org.hibernate.cfg.AnnotationConfiguration;
-import org.hibernate.SessionFactory;
-
-/**
- * Hibernate Utility class with a convenient method to get Session Factory
- * object.
- *
- * @author Kasia
- */
-public class Hibernate_Controller {
-
-    private static final SessionFactory sessionFactory;
-    
-    static {
-        try {
-            // Create the SessionFactory from standard (hibernate.cfg.xml) 
-            // config file.
-            sessionFactory = new AnnotationConfiguration().configure().buildSessionFactory();
-        } catch (Throwable ex) {
-            // Log the exception. 
-            System.err.println("Initial SessionFactory creation failed." + ex);
-            throw new ExceptionInInitializerError(ex);
-        }
-    }
-    
-    public static SessionFactory getSessionFactory() {
-        return sessionFactory;
-    }
-}
