+++ src/hibernate.cfg.xml
@@ -42,9 +42,11 @@ Contributor(s):
 <hibernate-configuration>
   <session-factory>
     <property name="hibernate.dialect">org.hibernate.dialect.MySQLDialect</property>
+    <property name="hibernate.connection.url">jdbc:mysql://localhost:3306/dziennik_elektroniczny</property>
     <property name="hibernate.connection.driver_class">com.mysql.jdbc.Driver</property>
-    <property name="hibernate.connection.url">jdbc:mysql://localhost:3306/dziennik_elektroniczny?zeroDateTimeBehavior=convertToNull</property>
     <property name="hibernate.connection.username">root</property>
+    <property name="hibernate.connection.password"/>
+   
     <mapping resource="mapping/Ocena.hbm.xml"/>
     <mapping resource="mapping/Uczen.hbm.xml"/>
     <mapping resource="mapping/Klasa.hbm.xml"/>
