+++ src/mapping/Nauczyciel.hbm.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <!DOCTYPE hibernate-mapping PUBLIC "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
 "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
-<!-- Generated 6 kwi 2019, 17:33:07 by Hibernate Tools 4.3.1 -->
+<!-- Generated 6 kwi 2019, 22:02:14 by Hibernate Tools 4.3.1 -->
 <hibernate-mapping>
     <class name="mapping.Nauczyciel" table="nauczyciel" catalog="dziennik_elektroniczny" optimistic-lock="version">
         <id name="pesel" type="int">
